@@ -1,218 +1,217 @@
-﻿namespace TNT.Drawing.Sample
+﻿namespace TNT.Drawing.Sample;
+
+partial class Form1
 {
-	partial class Form1
+	/// <summary>
+	/// Required designer variable.
+	/// </summary>
+	private System.ComponentModel.IContainer components = null;
+
+	/// <summary>
+	/// Clean up any resources being used.
+	/// </summary>
+	/// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
+	protected override void Dispose(bool disposing)
 	{
-		/// <summary>
-		/// Required designer variable.
-		/// </summary>
-		private System.ComponentModel.IContainer components = null;
-
-		/// <summary>
-		/// Clean up any resources being used.
-		/// </summary>
-		/// <param name="disposing">true if managed resources should be disposed; otherwise, false.</param>
-		protected override void Dispose(bool disposing)
+		if (disposing && (components != null))
 		{
-			if (disposing && (components != null))
-			{
-				components.Dispose();
-			}
-			base.Dispose(disposing);
+			components.Dispose();
 		}
+		base.Dispose(disposing);
+	}
 
-		#region Windows Form Designer generated code
+	#region Windows Form Designer generated code
 
-		/// <summary>
-		/// Required method for Designer support - do not modify
-		/// the contents of this method with the code editor.
-		/// </summary>
-		private void InitializeComponent()
-		{
-			this.splitContainer1 = new System.Windows.Forms.SplitContainer();
-			this.menuStrip1 = new System.Windows.Forms.MenuStrip();
-			this.menuToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
-			this.fitToolStripMenuItem1 = new System.Windows.Forms.ToolStripMenuItem();
-			this.saveToolStripMenuItem1 = new System.Windows.Forms.ToolStripMenuItem();
-			this.openToolStripMenuItem1 = new System.Windows.Forms.ToolStripMenuItem();
-			this.modeToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
-			this.lineToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
-			this.selectToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
-			this.layerToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
-			this.propertyGrid1 = new System.Windows.Forms.PropertyGrid();
-			this.statusStrip1 = new System.Windows.Forms.StatusStrip();
-			this.toolStripStatusLabel1 = new System.Windows.Forms.ToolStripStatusLabel();
-			this.alignToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
-			((System.ComponentModel.ISupportInitialize)(this.splitContainer1)).BeginInit();
-			this.splitContainer1.Panel1.SuspendLayout();
-			this.splitContainer1.Panel2.SuspendLayout();
-			this.splitContainer1.SuspendLayout();
-			this.menuStrip1.SuspendLayout();
-			this.statusStrip1.SuspendLayout();
-			this.SuspendLayout();
-			// 
-			// splitContainer1
-			// 
-			this.splitContainer1.Dock = System.Windows.Forms.DockStyle.Fill;
-			this.splitContainer1.Location = new System.Drawing.Point(0, 0);
-			this.splitContainer1.Name = "splitContainer1";
-			// 
-			// splitContainer1.Panel1
-			// 
-			this.splitContainer1.Panel1.Controls.Add(this.menuStrip1);
-			// 
-			// splitContainer1.Panel2
-			// 
-			this.splitContainer1.Panel2.Controls.Add(this.propertyGrid1);
-			this.splitContainer1.Panel2.Controls.Add(this.statusStrip1);
-			this.splitContainer1.Size = new System.Drawing.Size(1035, 688);
-			this.splitContainer1.SplitterDistance = 800;
-			this.splitContainer1.TabIndex = 2;
-			// 
-			// menuStrip1
-			// 
-			this.menuStrip1.Items.AddRange(new System.Windows.Forms.ToolStripItem[] {
+	/// <summary>
+	/// Required method for Designer support - do not modify
+	/// the contents of this method with the code editor.
+	/// </summary>
+	private void InitializeComponent()
+	{
+		this.splitContainer1 = new System.Windows.Forms.SplitContainer();
+		this.menuStrip1 = new System.Windows.Forms.MenuStrip();
+		this.menuToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+		this.fitToolStripMenuItem1 = new System.Windows.Forms.ToolStripMenuItem();
+		this.saveToolStripMenuItem1 = new System.Windows.Forms.ToolStripMenuItem();
+		this.openToolStripMenuItem1 = new System.Windows.Forms.ToolStripMenuItem();
+		this.modeToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+		this.lineToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+		this.selectToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+		this.layerToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+		this.propertyGrid1 = new System.Windows.Forms.PropertyGrid();
+		this.statusStrip1 = new System.Windows.Forms.StatusStrip();
+		this.toolStripStatusLabel1 = new System.Windows.Forms.ToolStripStatusLabel();
+		this.alignToolStripMenuItem = new System.Windows.Forms.ToolStripMenuItem();
+		((System.ComponentModel.ISupportInitialize)(this.splitContainer1)).BeginInit();
+		this.splitContainer1.Panel1.SuspendLayout();
+		this.splitContainer1.Panel2.SuspendLayout();
+		this.splitContainer1.SuspendLayout();
+		this.menuStrip1.SuspendLayout();
+		this.statusStrip1.SuspendLayout();
+		this.SuspendLayout();
+		// 
+		// splitContainer1
+		// 
+		this.splitContainer1.Dock = System.Windows.Forms.DockStyle.Fill;
+		this.splitContainer1.Location = new System.Drawing.Point(0, 0);
+		this.splitContainer1.Name = "splitContainer1";
+		// 
+		// splitContainer1.Panel1
+		// 
+		this.splitContainer1.Panel1.Controls.Add(this.menuStrip1);
+		// 
+		// splitContainer1.Panel2
+		// 
+		this.splitContainer1.Panel2.Controls.Add(this.propertyGrid1);
+		this.splitContainer1.Panel2.Controls.Add(this.statusStrip1);
+		this.splitContainer1.Size = new System.Drawing.Size(1035, 688);
+		this.splitContainer1.SplitterDistance = 800;
+		this.splitContainer1.TabIndex = 2;
+		// 
+		// menuStrip1
+		// 
+		this.menuStrip1.Items.AddRange(new System.Windows.Forms.ToolStripItem[] {
             this.menuToolStripMenuItem,
             this.modeToolStripMenuItem,
             this.layerToolStripMenuItem});
-			this.menuStrip1.Location = new System.Drawing.Point(0, 0);
-			this.menuStrip1.Name = "menuStrip1";
-			this.menuStrip1.Size = new System.Drawing.Size(800, 24);
-			this.menuStrip1.TabIndex = 0;
-			this.menuStrip1.Text = "menuStrip1";
-			// 
-			// menuToolStripMenuItem
-			// 
-			this.menuToolStripMenuItem.DropDownItems.AddRange(new System.Windows.Forms.ToolStripItem[] {
+		this.menuStrip1.Location = new System.Drawing.Point(0, 0);
+		this.menuStrip1.Name = "menuStrip1";
+		this.menuStrip1.Size = new System.Drawing.Size(800, 24);
+		this.menuStrip1.TabIndex = 0;
+		this.menuStrip1.Text = "menuStrip1";
+		// 
+		// menuToolStripMenuItem
+		// 
+		this.menuToolStripMenuItem.DropDownItems.AddRange(new System.Windows.Forms.ToolStripItem[] {
             this.fitToolStripMenuItem1,
             this.saveToolStripMenuItem1,
             this.openToolStripMenuItem1,
             this.alignToolStripMenuItem});
-			this.menuToolStripMenuItem.Name = "menuToolStripMenuItem";
-			this.menuToolStripMenuItem.Size = new System.Drawing.Size(50, 20);
-			this.menuToolStripMenuItem.Text = "Menu";
-			// 
-			// fitToolStripMenuItem1
-			// 
-			this.fitToolStripMenuItem1.Name = "fitToolStripMenuItem1";
-			this.fitToolStripMenuItem1.Size = new System.Drawing.Size(180, 22);
-			this.fitToolStripMenuItem1.Text = "Fit";
-			this.fitToolStripMenuItem1.Click += new System.EventHandler(this.fitToolStripMenuItem_Click);
-			// 
-			// saveToolStripMenuItem1
-			// 
-			this.saveToolStripMenuItem1.Name = "saveToolStripMenuItem1";
-			this.saveToolStripMenuItem1.Size = new System.Drawing.Size(180, 22);
-			this.saveToolStripMenuItem1.Text = "Save";
-			this.saveToolStripMenuItem1.Click += new System.EventHandler(this.saveToolStripMenuItem_Click);
-			// 
-			// openToolStripMenuItem1
-			// 
-			this.openToolStripMenuItem1.Name = "openToolStripMenuItem1";
-			this.openToolStripMenuItem1.Size = new System.Drawing.Size(180, 22);
-			this.openToolStripMenuItem1.Text = "Open";
-			this.openToolStripMenuItem1.Click += new System.EventHandler(this.openToolStripMenuItem_Click);
-			// 
-			// modeToolStripMenuItem
-			// 
-			this.modeToolStripMenuItem.DropDownItems.AddRange(new System.Windows.Forms.ToolStripItem[] {
+		this.menuToolStripMenuItem.Name = "menuToolStripMenuItem";
+		this.menuToolStripMenuItem.Size = new System.Drawing.Size(50, 20);
+		this.menuToolStripMenuItem.Text = "Menu";
+		// 
+		// fitToolStripMenuItem1
+		// 
+		this.fitToolStripMenuItem1.Name = "fitToolStripMenuItem1";
+		this.fitToolStripMenuItem1.Size = new System.Drawing.Size(180, 22);
+		this.fitToolStripMenuItem1.Text = "Fit";
+		this.fitToolStripMenuItem1.Click += new System.EventHandler(this.fitToolStripMenuItem_Click);
+		// 
+		// saveToolStripMenuItem1
+		// 
+		this.saveToolStripMenuItem1.Name = "saveToolStripMenuItem1";
+		this.saveToolStripMenuItem1.Size = new System.Drawing.Size(180, 22);
+		this.saveToolStripMenuItem1.Text = "Save";
+		this.saveToolStripMenuItem1.Click += new System.EventHandler(this.saveToolStripMenuItem_Click);
+		// 
+		// openToolStripMenuItem1
+		// 
+		this.openToolStripMenuItem1.Name = "openToolStripMenuItem1";
+		this.openToolStripMenuItem1.Size = new System.Drawing.Size(180, 22);
+		this.openToolStripMenuItem1.Text = "Open";
+		this.openToolStripMenuItem1.Click += new System.EventHandler(this.openToolStripMenuItem_Click);
+		// 
+		// modeToolStripMenuItem
+		// 
+		this.modeToolStripMenuItem.DropDownItems.AddRange(new System.Windows.Forms.ToolStripItem[] {
             this.lineToolStripMenuItem,
             this.selectToolStripMenuItem});
-			this.modeToolStripMenuItem.Name = "modeToolStripMenuItem";
-			this.modeToolStripMenuItem.Size = new System.Drawing.Size(50, 20);
-			this.modeToolStripMenuItem.Text = "Mode";
-			// 
-			// lineToolStripMenuItem
-			// 
-			this.lineToolStripMenuItem.Name = "lineToolStripMenuItem";
-			this.lineToolStripMenuItem.Size = new System.Drawing.Size(105, 22);
-			this.lineToolStripMenuItem.Text = "Line";
-			this.lineToolStripMenuItem.Click += new System.EventHandler(this.lineToolStripMenuItem_Click);
-			// 
-			// selectToolStripMenuItem
-			// 
-			this.selectToolStripMenuItem.Name = "selectToolStripMenuItem";
-			this.selectToolStripMenuItem.Size = new System.Drawing.Size(105, 22);
-			this.selectToolStripMenuItem.Text = "Select";
-			this.selectToolStripMenuItem.Click += new System.EventHandler(this.lineToolStripMenuItem_Click);
-			// 
-			// layerToolStripMenuItem
-			// 
-			this.layerToolStripMenuItem.Name = "layerToolStripMenuItem";
-			this.layerToolStripMenuItem.Size = new System.Drawing.Size(47, 20);
-			this.layerToolStripMenuItem.Text = "Layer";
-			// 
-			// propertyGrid1
-			// 
-			this.propertyGrid1.Dock = System.Windows.Forms.DockStyle.Fill;
-			this.propertyGrid1.Location = new System.Drawing.Point(0, 0);
-			this.propertyGrid1.Name = "propertyGrid1";
-			this.propertyGrid1.Size = new System.Drawing.Size(231, 666);
-			this.propertyGrid1.TabIndex = 3;
-			this.propertyGrid1.PropertyValueChanged += new System.Windows.Forms.PropertyValueChangedEventHandler(this.propertyGrid1_PropertyValueChanged);
-			// 
-			// statusStrip1
-			// 
-			this.statusStrip1.Items.AddRange(new System.Windows.Forms.ToolStripItem[] {
+		this.modeToolStripMenuItem.Name = "modeToolStripMenuItem";
+		this.modeToolStripMenuItem.Size = new System.Drawing.Size(50, 20);
+		this.modeToolStripMenuItem.Text = "Mode";
+		// 
+		// lineToolStripMenuItem
+		// 
+		this.lineToolStripMenuItem.Name = "lineToolStripMenuItem";
+		this.lineToolStripMenuItem.Size = new System.Drawing.Size(105, 22);
+		this.lineToolStripMenuItem.Text = "Line";
+		this.lineToolStripMenuItem.Click += new System.EventHandler(this.lineToolStripMenuItem_Click);
+		// 
+		// selectToolStripMenuItem
+		// 
+		this.selectToolStripMenuItem.Name = "selectToolStripMenuItem";
+		this.selectToolStripMenuItem.Size = new System.Drawing.Size(105, 22);
+		this.selectToolStripMenuItem.Text = "Select";
+		this.selectToolStripMenuItem.Click += new System.EventHandler(this.lineToolStripMenuItem_Click);
+		// 
+		// layerToolStripMenuItem
+		// 
+		this.layerToolStripMenuItem.Name = "layerToolStripMenuItem";
+		this.layerToolStripMenuItem.Size = new System.Drawing.Size(47, 20);
+		this.layerToolStripMenuItem.Text = "Layer";
+		// 
+		// propertyGrid1
+		// 
+		this.propertyGrid1.Dock = System.Windows.Forms.DockStyle.Fill;
+		this.propertyGrid1.Location = new System.Drawing.Point(0, 0);
+		this.propertyGrid1.Name = "propertyGrid1";
+		this.propertyGrid1.Size = new System.Drawing.Size(231, 666);
+		this.propertyGrid1.TabIndex = 3;
+		this.propertyGrid1.PropertyValueChanged += new System.Windows.Forms.PropertyValueChangedEventHandler(this.propertyGrid1_PropertyValueChanged);
+		// 
+		// statusStrip1
+		// 
+		this.statusStrip1.Items.AddRange(new System.Windows.Forms.ToolStripItem[] {
             this.toolStripStatusLabel1});
-			this.statusStrip1.Location = new System.Drawing.Point(0, 666);
-			this.statusStrip1.Name = "statusStrip1";
-			this.statusStrip1.Size = new System.Drawing.Size(231, 22);
-			this.statusStrip1.TabIndex = 2;
-			this.statusStrip1.Text = "statusStrip1";
-			// 
-			// toolStripStatusLabel1
-			// 
-			this.toolStripStatusLabel1.Name = "toolStripStatusLabel1";
-			this.toolStripStatusLabel1.Size = new System.Drawing.Size(118, 17);
-			this.toolStripStatusLabel1.Text = "toolStripStatusLabel1";
-			// 
-			// alignToolStripMenuItem
-			// 
-			this.alignToolStripMenuItem.Name = "alignToolStripMenuItem";
-			this.alignToolStripMenuItem.Size = new System.Drawing.Size(180, 22);
-			this.alignToolStripMenuItem.Text = "Align";
-			this.alignToolStripMenuItem.Click += new System.EventHandler(this.alignToolStripMenuItem_Click);
-			// 
-			// Form1
-			// 
-			this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
-			this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-			this.AutoScroll = true;
-			this.ClientSize = new System.Drawing.Size(1035, 688);
-			this.Controls.Add(this.splitContainer1);
-			this.MainMenuStrip = this.menuStrip1;
-			this.Name = "Form1";
-			this.Text = "Form1";
-			this.splitContainer1.Panel1.ResumeLayout(false);
-			this.splitContainer1.Panel1.PerformLayout();
-			this.splitContainer1.Panel2.ResumeLayout(false);
-			this.splitContainer1.Panel2.PerformLayout();
-			((System.ComponentModel.ISupportInitialize)(this.splitContainer1)).EndInit();
-			this.splitContainer1.ResumeLayout(false);
-			this.menuStrip1.ResumeLayout(false);
-			this.menuStrip1.PerformLayout();
-			this.statusStrip1.ResumeLayout(false);
-			this.statusStrip1.PerformLayout();
-			this.ResumeLayout(false);
+		this.statusStrip1.Location = new System.Drawing.Point(0, 666);
+		this.statusStrip1.Name = "statusStrip1";
+		this.statusStrip1.Size = new System.Drawing.Size(231, 22);
+		this.statusStrip1.TabIndex = 2;
+		this.statusStrip1.Text = "statusStrip1";
+		// 
+		// toolStripStatusLabel1
+		// 
+		this.toolStripStatusLabel1.Name = "toolStripStatusLabel1";
+		this.toolStripStatusLabel1.Size = new System.Drawing.Size(118, 17);
+		this.toolStripStatusLabel1.Text = "toolStripStatusLabel1";
+		// 
+		// alignToolStripMenuItem
+		// 
+		this.alignToolStripMenuItem.Name = "alignToolStripMenuItem";
+		this.alignToolStripMenuItem.Size = new System.Drawing.Size(180, 22);
+		this.alignToolStripMenuItem.Text = "Align";
+		this.alignToolStripMenuItem.Click += new System.EventHandler(this.alignToolStripMenuItem_Click);
+		// 
+		// Form1
+		// 
+		this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
+		this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
+		this.AutoScroll = true;
+		this.ClientSize = new System.Drawing.Size(1035, 688);
+		this.Controls.Add(this.splitContainer1);
+		this.MainMenuStrip = this.menuStrip1;
+		this.Name = "Form1";
+		this.Text = "Form1";
+		this.splitContainer1.Panel1.ResumeLayout(false);
+		this.splitContainer1.Panel1.PerformLayout();
+		this.splitContainer1.Panel2.ResumeLayout(false);
+		this.splitContainer1.Panel2.PerformLayout();
+		((System.ComponentModel.ISupportInitialize)(this.splitContainer1)).EndInit();
+		this.splitContainer1.ResumeLayout(false);
+		this.menuStrip1.ResumeLayout(false);
+		this.menuStrip1.PerformLayout();
+		this.statusStrip1.ResumeLayout(false);
+		this.statusStrip1.PerformLayout();
+		this.ResumeLayout(false);
 
-		}
-
-		#endregion
-
-		private System.Windows.Forms.SplitContainer splitContainer1;
-		private System.Windows.Forms.StatusStrip statusStrip1;
-		private System.Windows.Forms.ToolStripStatusLabel toolStripStatusLabel1;
-		private System.Windows.Forms.PropertyGrid propertyGrid1;
-		private System.Windows.Forms.MenuStrip menuStrip1;
-		private System.Windows.Forms.ToolStripMenuItem menuToolStripMenuItem;
-		private System.Windows.Forms.ToolStripMenuItem fitToolStripMenuItem1;
-		private System.Windows.Forms.ToolStripMenuItem saveToolStripMenuItem1;
-		private System.Windows.Forms.ToolStripMenuItem openToolStripMenuItem1;
-		private System.Windows.Forms.ToolStripMenuItem modeToolStripMenuItem;
-		private System.Windows.Forms.ToolStripMenuItem lineToolStripMenuItem;
-		private System.Windows.Forms.ToolStripMenuItem selectToolStripMenuItem;
-		private System.Windows.Forms.ToolStripMenuItem layerToolStripMenuItem;
-		private System.Windows.Forms.ToolStripMenuItem alignToolStripMenuItem;
 	}
+
+	#endregion
+
+	private System.Windows.Forms.SplitContainer splitContainer1;
+	private System.Windows.Forms.StatusStrip statusStrip1;
+	private System.Windows.Forms.ToolStripStatusLabel toolStripStatusLabel1;
+	private System.Windows.Forms.PropertyGrid propertyGrid1;
+	private System.Windows.Forms.MenuStrip menuStrip1;
+	private System.Windows.Forms.ToolStripMenuItem menuToolStripMenuItem;
+	private System.Windows.Forms.ToolStripMenuItem fitToolStripMenuItem1;
+	private System.Windows.Forms.ToolStripMenuItem saveToolStripMenuItem1;
+	private System.Windows.Forms.ToolStripMenuItem openToolStripMenuItem1;
+	private System.Windows.Forms.ToolStripMenuItem modeToolStripMenuItem;
+	private System.Windows.Forms.ToolStripMenuItem lineToolStripMenuItem;
+	private System.Windows.Forms.ToolStripMenuItem selectToolStripMenuItem;
+	private System.Windows.Forms.ToolStripMenuItem layerToolStripMenuItem;
+	private System.Windows.Forms.ToolStripMenuItem alignToolStripMenuItem;
 }
 
