@@ -1,19 +1,16 @@
-﻿using System;
-using System.Windows.Forms;
+namespace TNT.Drawing.Sample;
 
-namespace TNT.Drawing.Sample
+internal static class Program
 {
-	static class Program
-	{
-		/// <summary>
-		/// The main entry point for the application.
-		/// </summary>
-		[STAThread]
-		static void Main()
-		{
-			Application.EnableVisualStyles();
-			Application.SetCompatibleTextRenderingDefault(false);
-			Application.Run(new Form1());
-		}
-	}
+  /// <summary>
+  ///  The main entry point for the application.
+  /// </summary>
+  [STAThread]
+  static void Main()
+  {
+    // To customize application configuration such as set high DPI settings or default font,
+    // see https://aka.ms/applicationconfiguration.
+    ApplicationConfiguration.Initialize();
+    Application.Run(new Form1());
+  }
 }
